+++ demos/cyto/src/commonMain/kotlin/org/emerge/demo/cyto/sim/systems/CytoContactSystem.kt
@@ -59,7 +59,7 @@ object CytoContactSystem : EcsSystem<CytoConfig, SimState, CytoInput> {
             builder.update<ImpulseComponent>(a) { ImpulseComponent(vel = normal * (push * weightA)) + it }
             builder.update<ImpulseComponent>(b) { ImpulseComponent(vel = -(normal * (push * weightB))) + it }
 
-            val touchAmount = contact.penetration.toFloat()
+            val touchAmount = contact.penetration
             builder.update<CytoCellComponent>(a) { c -> (c ?: cellA).let { it.copy(touch = it.touch + touchAmount) } }
             builder.update<CytoCellComponent>(b) { c -> (c ?: cellB).let { it.copy(touch = it.touch + touchAmount) } }
         }
